+++ src/merge/mkvmerge.cpp
@@ -139,6 +139,8 @@ set_usage() {
   usage_text += Y("  --disable-lacing         Do not use lacing.\n");
   usage_text += Y("  --disable-track-statistics-tags\n"
                   "                           Do not write tags with track statistics.\n");
+  usage_text += Y("  --disable-language-ietf  Do not write LanguageIETF track header and\n"
+                  "                           ChapLanguageIETF chapter elements.\n");
   usage_text +=   "\n";
   usage_text += Y(" File splitting, linking, appending and concatenating (more global options):\n");
   usage_text += Y("  --split <d[K,M,G]|HH:MM:SS|s>\n"
@@ -2397,6 +2399,9 @@ parse_args(std::vector<std::string> args) {
 
     else if (this_arg == "--disable-track-statistics-tags")
       g_no_track_statistics_tags = true;
+
+    else if (this_arg == "--disable-language-ietf")
+      mtx::bcp47::language_c::disable();
 
     else if (this_arg == "--attachment-description") {
       if (no_next_arg)